+++ esmodules/message-hooks.js
@@ -1,48 +1,10 @@
-import { getChatTheme, getDefaultColorScheme } from "./ui-theme.js";
-
-const rgb2hex = (rgb) =>
-  `#${rgb
-    .match(/^rgb\((\d+),\s*(\d+),\s*(\d+)\)$/)
-    .slice(1)
-    .map((n) => parseInt(n, 10).toString(16).padStart(2, "0"))
-    .join("")}`;
+import { getChatTheme, getDefaultColorScheme, lookupThemeAndSchemeForKey } from "./ui-theme.js";
 
 Hooks.on("renderChatMessage", (chatMessage, html, messageData) => {
-  html[0].dataset.colorScheme = getMessageColorScheme(chatMessage, html, messageData);
   if (html[0].hasAttribute("style")) {
     html.css("border-color", "");
   }
 });
-
-function getMessageColorScheme(chatMessage, html, messageData) {
-  const chatColorSchemeSetting = game.settings.get("pf2e-dorako-ui", "theme.chat-message-color-scheme");
-  const chatTheme = game.settings.get("pf2e-dorako-ui", "theme.chat-message-theme");
-  const defaultColorScheme = getDefaultColorScheme(chatTheme);
-  let colorScheme = null;
-  switch (chatColorSchemeSetting) {
-    case "default":
-      colorScheme = defaultColorScheme;
-      break;
-    case "prefer-dark":
-      colorScheme = "dark";
-      break;
-    case "prefer-light":
-      colorScheme = "light";
-      break;
-    case "alliance":
-      colorScheme = defaultColorScheme; // Fallback for GM speaking from non-actors
-      if (!chatMessage?.actor?.alliance) {
-        colorScheme = defaultColorScheme;
-      } else {
-        colorScheme = chatMessage.actor.alliance === "opposition" ? "dark" : "light";
-      }
-      break;
-    case "gm-vs-players":
-      colorScheme = chatMessage.user.isGM ? "dark" : "light";
-      break;
-  }
-  return colorScheme;
-}
 
 // Chat cards
 Hooks.on("renderChatMessage", (chatMessage, html, messageData) => {
@@ -56,30 +18,31 @@ Hooks.on("renderChatMessage", (chatMessage, html, messageData) => {
 
   let html0 = html[0];
   html0.style.setProperty("--player-color", chatMessage?.user?.color ?? "#DAC0FB");
-  html0.dataset.headerTextColorScheme = calcHeaderTextColor(html, chatMessage);
 
-  const uiTheme = getChatTheme();
-  const { dorakoUiTheme, colorScheme } = uiTheme;
-  if (uiTheme) {
-    html0.dataset.dorakoUiTheme = dorakoUiTheme;
-    html0.dataset.dorakoUiScope = "unlimited";
+  const isSecretDisposition =
+    game?.scenes?.get(chatMessage?.speaker?.scene)?.tokens?.get(chatMessage?.speaker?.token)?.disposition == -2;
+  const { dorakoUiTheme, colorScheme } =
+    chatMessage?.actor?.alliance === "opposition" && !isSecretDisposition
+      ? lookupThemeAndSchemeForKey(game.settings.get("pf2e-dorako-ui", "theme.chat-message-opposition-theme"))
+      : lookupThemeAndSchemeForKey(game.settings.get("pf2e-dorako-ui", "theme.chat-message-standard-theme"));
+  html0.dataset.dorakoUiTheme = dorakoUiTheme;
+  html0.dataset.colorScheme = colorScheme;
+  html0.dataset.dorakoUiScope = "unlimited";
+
+  const headerStyle = game.settings.get("pf2e-dorako-ui", "theme.chat-message-header-style");
+  if (dorakoUiTheme === "crb" && headerStyle !== "none") {
+    html0.dataset.hasHeader = "";
   }
-
   themeHeader(html, chatMessage);
 });
 
 function themeHeader(html, message) {
   let messageHeader = html.find(".message-header")[0];
+  const headerColor = getHeaderColor(html, message);
+  messageHeader.style.setProperty("--header-color", headerColor);
 
-  const headerStyle = game.settings.get("pf2e-dorako-ui", "theme.chat-message-header-style");
-  if (headerStyle != "none") {
-    let bgCol = getHeaderColor(html, message);
-    messageHeader.style.setProperty("--header-color", bgCol);
-    messageHeader.classList.add("dorako-header");
-  }
-
-  let textColTheme = calcHeaderTextColor(html, message);
-  messageHeader.classList.add(textColTheme);
+  let textColTheme = calcHeaderTextColor(headerColor);
+  html[0].dataset.headerTextColorScheme = textColTheme;
 
   // some modules add different timestamps and hide the original, like dfce-simple-timestamp
   let time = html.find("time")[0];
@@ -108,26 +71,10 @@ function invertColor(color) {
   return color === "dark" ? "light" : "dark";
 }
 
-function calcHeaderTextColor(html, message) {
-  const headerStyle = game.settings.get("pf2e-dorako-ui", "theme.chat-message-header-style");
-  const chatTheme = game.settings.get("pf2e-dorako-ui", "theme.chat-message-theme");
-  // const defaultColorScheme = getDefaultColorScheme(chatTheme);
-  const inverseColorScheme = invertColor(getMessageColorScheme(message, html, null));
-  // const inverse = invertColor(defaultColorScheme);
-  const messageHeader = html.find(".message-header")[0];
-  if (headerStyle === "none" && chatTheme == "crb") {
-    return inverseColorScheme;
-  } else if (headerStyle === "none") {
-    return "light";
-  }
-
-  // let bgCol = messageHeader.style.backgroundColor;
-
-  let bgCol = getHeaderColor(html, message);
-  // bgCol = rgb2hex(bgCol);
-  var r = parseInt(bgCol.substr(1, 2), 16);
-  var g = parseInt(bgCol.substr(3, 2), 16);
-  var b = parseInt(bgCol.substr(5, 2), 16);
+function calcHeaderTextColor(headerColor) {
+  var r = parseInt(headerColor.substr(1, 2), 16);
+  var g = parseInt(headerColor.substr(3, 2), 16);
+  var b = parseInt(headerColor.substr(5, 2), 16);
   var yiq = (r * 299 + g * 587 + b * 114) / 1000;
 
   if (yiq >= 180) {